+++ src/lib.rs
@@ -7,10 +7,10 @@
     unreachable_pub
 )]
 #![doc(
-    html_logo_url = "https://raw.githubusercontent.com/AaronErhardt/relm4/main/assets/Relm_logo.svg"
+    html_logo_url = "https://raw.githubusercontent.com/https://raw.githubusercontent.com/Relm4/relm4//relm4/main/assets/Relm_logo.svg"
 )]
 #![doc(
-    html_favicon_url = "https://raw.githubusercontent.com/AaronErhardt/relm4/main/assets/Relm_logo.svg"
+    html_favicon_url = "https://raw.githubusercontent.com/https://raw.githubusercontent.com/Relm4/relm4//relm4/main/assets/Relm_logo.svg"
 )]
 
 pub mod actions;