+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
++ relm4-macros: Some doc link fixes
+
 ## 0.4.3 - 2022-3-13 
 
 ### Added